refactor: 폴더 구조 정리 및 관련 파일 이동 #46
