+++ fe/src/components/analysis/AnalysisView.tsx
@@ -6,8 +6,8 @@ import { ANALYSIS_CONFIG } from '@/constants/analysis';
 import AnalysisEmpty from '@/components/analysis/common/AnalysisEmpty';
 import AnalysisLoading from '@/components/analysis/common/AnalysisLoading';
 import AnalysisSection from '@/components/analysis/common/AnalysisSection';
-import CategoryAnalysisList from '@/components/analysis/common/CategoryAnalysisList';
-import CategoryChart from '@/components/analysis/common/CategoryChart';
+import CategoryAnalysisList from '@/components/analysis/CategoryAnalysisList';
+import CategoryChart from '@/components/analysis/CategoryChart';
 import { PieChart } from 'lucide-react';
 import { Separator } from '@/components/ui/separator';
 import { TransactionType } from '@/types/analysis';
