+++ fe/src/components/analysis/CategoryAnalysisList.tsx
