+++ fe/src/components/analysis/CategoryChart.tsx
